- Modified the wrf_grid function Basically, I had (wrongly) assumed in the past that the centre of the grid had to be coincident with the centre of the projection for this function to work properly. When looking at it again, I realized this may well not be the case and just taken out the statement that caused the function to bail out if the above condition was not met. Well, I have not tested the functionality extensively, nevertheless, it seems to work fine without further modification.
- Added the folder "borrowed_code" in trunk
The idea is to use this folder for all the compiled pieces of code that we
"borrow" from wherever. In this case, I have ripped three fortran functions
out of the wrf_to_grads code that ships as part of ARWPost (I believe... it's
been some time...). The one that I cared about in the first place was the
function that computes the sea pressure level atmospheric pressure. This is
the one that I used in the following new function.

- Added a function to compute the sea level pressure
NB: I am dropping the adjective mean (as in mslp) because there is no 
averaging of the temperature for the (isothermal) virtual layer  of air
assumed to exist below the surface.
This function basically takes "raw" arrays out of a wrfout files and returns
the sea level pressure. It works both with 3D and 4D array by assuming the
ordering of the dimensions is COARDS -> foo(time, lvl, lat, lon). 
NB: I've had problems of pynio running out of memory for very big grids ->
451x331x50x31 (nx x ny x nz x nt) on a 4Gb RAM machine. It may be needed to
divide the arrays up in tiles at run time to reduce the memory footprint.
Alternatively, we could write a quick script that only calculates the slp and
dumps in some auxiliary file from which it is then retrieved for plotting...
for now I'll just live with it.